[Weetbix] Revise Weetbix ingestion to allow cross-project joins.
CV runs in one LUCI Project may include builds from other LUCI
projects. This was not considered in the original ingestion
implementation.

Revise the implementation so that this is possible:
- Build ID is now used as the join key, not (Project, Build ID).
- As before, data is always ingested into the project that the
  build is part of (technically, the project of the invocation
  being ingested), not the project of the CV run.
- CV Runs are always processed, regardless of if their project is
  configured, as they may contain builds from other projects which
  are configured.
- Metrics are updated to measure join performance from both
  directions (Build to CV, CV to Build).

As part of this change, the IngestionControl table will be
replaced with a new table, called Ingestions. First the new table
will be created, then the new release will be deployed, and then
the old table will be deleted. This will prevent rollback.

The migration will result in the loss of data for builds which are
pending join at time of the migration. It is considered the value
of this data is not significant compared to the effort of designing
a process to migrate the data.

BUG=1243174
TEST=go test infra/appengine/weetbix/...

Change-Id: Ia1a7c4cc22db6059e68afa6fe948df37f5ea9e89
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3543014
Reviewed-by: Matthew Warton <mwarton@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>
Cr-Commit-Position: refs/heads/main@{#47862}