[bcid-for-cipd] Add provenance for 3pp packages.
3pp packages use recipe to attach provenance today, this is changing in
this CL as provenance generation need to be isolated from recipe itself.
This is an L2 requirement that we are going to solve with Snoopy a
process inspection tool running on each machine.

At the time of the CL, snoopy is distributed to only Linux machine via
puppet and running on localhost:11000. Technically there isn't any
blocker for distributing it to other platforms but to closely monitor
roll out we have chosen Linux to be pilot.

Similar roll out is happening now on Chrome recipe (Google-internal).

Bug: 1093088
Change-Id: Ib421b90279565487187a48bbc6635a6448a30a75
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3554717
Reviewed-by: Brian Ryner <bryner@google.com>
Commit-Queue: Akash Mukherjee <akashmukherjee@google.com>
Cr-Commit-Position: refs/heads/main@{#47974}