Remove orodley@google.com from appengine/chromium_status_py3/OWNERS
This is an automatically generated changelist based on group memberships and affiliations, with some transition delay.

To approve this change, CR+1/2 this change and submit. To reject this change, CR -1 this change, and the bot will abandon the change.

To report an issue, file a bug in the Infra>Security component.

Change-Id: If0e8285e7b7f1c1be180735aa48a05dc1c71cf1d
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3555653
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Tuan Nguyen <nqmtuan@google.com>
Commit-Queue: Tuan Nguyen <nqmtuan@google.com>
Cr-Commit-Position: refs/heads/main@{#47984}