[karte] Don't perform repeated tests on explicit `make test` and `make check`.
Repeated tests have value for detecting non-determinism (which can be introduced by,
for examle, iterating over a map), but they also produce a lot of output on test failure.

It makes sense to run tests repeatedly on the default target for make so that devs that use
`make` will see these sorts of errors if they occur.

However, when running `make test` or `make check`, we frequently want to inspect the reasons why
a specific test failed and it's important to keep the output concise.

BUG=b:222358824
TEST=make test

Change-Id: I07529255680a4990b738f58f5a8da3ce3cc77187
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3556793
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Cr-Commit-Position: refs/heads/main@{#48005}