recovery: Add "servodtool" check in the process of servod start.
Currently, the process of servod start is lacking the ability to check
whether the servod process is really being started and usable.

Therefore, this cl introduces the "servodtool" check immediately after
we run the servod start command in order to make sure the started servod
is usable.

BUG=b:226950953
TEST=manual test

Change-Id: I4b1538691b1533fbad23c1859bd8b7ee7933cf94
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3551720
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Yunzhi Yu <yunzhiyu@google.com>
Cr-Commit-Position: refs/heads/main@{#48024}