[recovery] -- Reopen tree.
As a potential mitigation for b:227523207, do not run any of the retry tests in parallel.
If this also results in flakiness, the next mitigation is to disable the test entirely.

Subsequent CLs will alter this test to have time limits that are less likely to flake in the
CI environment.

BUG=b:227523207
TEST=make test

Change-Id: I7d985cd9b2c94dc1bc6795eff8519b4527fc77fe
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3561149
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Cr-Commit-Position: refs/heads/main@{#48067}