recovery: Make Servo Fail state correct in servo config file.
Currently, the order of the critical actions for triggering the set of
the bad Servo state is different than python. In order to set the
correct bad servo state, needs to adjust the order of the critical
actions as well as some dependencies according to labpack.

Here is the order specified by labpack to set the bad servo state:
https://source.corp.google.com/chromeos_public/src/third_party/labpack/files/server/hosts/servo_host.py;l=1560

BUG=b:222358559
TEST=manual test

Change-Id: Iddd5298909706ed128227f55a422273da5519f2e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3561989
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Yunzhi Yu <yunzhiyu@google.com>
Cr-Commit-Position: refs/heads/main@{#48169}