Use the path argument to DownloadDiff in chromium bootstrapper.
The path argument to DownloadDiff limits the patch to just the specified
path, which will be empty if the path is not changed in the diff.
Whether or not the file is affected can be determined by whether the
returned diff is empty or not. This allows simplifying the code by
removing the use as well as the fake implementation of the ListFiles
RPC.

Change-Id: Icc6344a137ef03bf8d9e304e34521cceba1b49f5
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3581410
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Gary Tong <gatong@chromium.org>
Commit-Queue: Gary Tong <gatong@chromium.org>
Cr-Commit-Position: refs/heads/main@{#48318}