Take a base argument in DownloadDiff in chromium bootstrapper.
In order for the chromium bootstrapper to detect if a change to a DEPS
file results in a change to the bootstrapped properties when using the
dependency_project field in $bootstrap/properties, it will be necessary
to be able to get the diff of the properties file between two arbitrary
revisions. This change adds support for specifying the base revision and
a PARENT constant that can be passed for the base argument to get the
existing diff-against-parent behavior.

Bug: 1313607
Change-Id: I4625e7dec0911bbfdf0f7797935ec8a452ac0b6b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3582547
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Gary Tong <gatong@chromium.org>
Commit-Queue: Gary Tong <gatong@chromium.org>
Cr-Commit-Position: refs/heads/main@{#48342}