Update the gitiles fake to return the log for a revision.
Currently, it is assumed that the committish value in the LogRequest
will be a ref, but for supporting dependency projects requests will be
sent where the Committish value may be a ref or a revision. This change
updates it to return the log for a revision if the Committish is present
in Revisions and not in Refs.

Bug: 1313607
Change-Id: I9938fc36f9a009f271e7a5b6af57642cecd4ea6a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3582866
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Gary Tong <gatong@chromium.org>
Commit-Queue: Gary Tong <gatong@chromium.org>
Cr-Commit-Position: refs/heads/main@{#48371}