[Weetbix] Add metric to trace ingestion outcome by build project.
Logging indicates a lot of builds are being discarded by Weetbix
because of no access. This revealed that the proportion of builds
being ignored by ingestion is currently not being tracked in any
metrics -- both ignored and success looks like a success as far
as monitoring is concerned.

Add a metric so that we have more visibility into what is going on.

Re-order checks so that builds for which projects are not
configured in Weetbix are discarded earlier. This may need to
be revisited in future as test history will likely want data
for all projects, so access issues will need to be resolved.
This will clean up the logs.

BUG=1243174
TEST=INTEGRATION_TESTS=1 go test infra/appengine/weetbix/...

Change-Id: Ic5441df579807f66aabd9f96bf4fa9332eb661ee
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3586567
Reviewed-by: Matthew Warton <mwarton@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>
Cr-Commit-Position: refs/heads/main@{#48425}