[monorail] Remove dependency on script-ext-html-webpack-plugin.
The package is no longer maintained and doesn't support Webpack 5.
We want to be on Webpack 5 to resolve our npm audit.

script-ext-html-webpack-plugin is used to add the `nonce` and `type`
attributes to the script tags generated by html-webpack-plugin, but
we can use templates to achieve the same result.

Bug: monorail:10502
Change-Id: I40754d349935d7482c27317bcd6d55ba388ff8be
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3597716
Auto-Submit: Dave Tu <dtu@chromium.org>
Reviewed-by: Andrew Chang <andrewjc@google.com>
Commit-Queue: Andrew Chang <andrewjc@google.com>
Cr-Commit-Position: refs/heads/main@{#48467}